+++ internal/provider/resource_docker_volume.go
@@ -64,41 +64,6 @@ func resourceDockerVolume() *schema.Resource {
 	}
 }
 
-func resourceDockerVolumeV0() *schema.Resource {
-	return &schema.Resource{
-		// This is only used for state migration, so the CRUD
-		// callbacks are no longer relevant
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Type:     schema.TypeString,
-				Optional: true,
-				Computed: true,
-				ForceNew: true,
-			},
-			"labels": {
-				Type:     schema.TypeMap,
-				Optional: true,
-				ForceNew: true,
-			},
-			"driver": {
-				Type:     schema.TypeString,
-				Optional: true,
-				Computed: true,
-				ForceNew: true,
-			},
-			"driver_opts": {
-				Type:     schema.TypeMap,
-				Optional: true,
-				ForceNew: true,
-			},
-			"mountpoint": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-		},
-	}
-}
-
 func resourceDockerVolumeCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*ProviderConfig).DockerClient
